Extract methods for checking links from Article into new class.
The #set_link_checker method does not need to be inside Goldencobra::Article. It is currently used only inside a rake task. This was changed to use the method from a new class Goldencobra::LinkChecker. The goal should be to keep the Goldencobra::Article class as focused as possible (Single responsibility principle).