Articles of kind "INDEX" display their children
If a new article of kind "index" was created, the initial value for article_for_index_id was nil. Most users forgot the set this one afterwards though. The nature of the select list in the backend interface made it so, that if no specific value was set, the id of the startpage article was set.

This change makes sure that the article gets its own ID as article_for_index_id if no value is given on #create.

GC-149