Backport for "Highlights parts of active navigation path"
The ActivePath for highlighting the navigation was only available if the whole
url path matched a possible Goldencobra::Menue entry. This meant that there had
to be unnecessary Goldencobra::Menue records just for highlighting the
navigation.
This change makes it so that if no matched Goldencobra::Menue entry is found,
the path is cropped of the last "/" part. That means that if
"/der-tagesspiegel/klassische-anzeigenformen/eckfelder" isn't matched, the next
lookup is for "/der-tagesspiegel/klassische-anzeigenformen". If a match is
found, the ids for the remaining parts are returned.
A possible side effect is that the lookup for Goldencobra::Menue entries that
match the current path might take longer. The positive side effect is, that
there are more times that an active path array is returned.
JIRA-TASK: QB-457