Gracefully handle 404 responses
Bots, maliciously acting people and many other 'things' request routes
in our Rails apps, that are simply not supported. Because we have a
'catch-all' route in Golden Cobra, we have to make sure these requests
aren't routed to existing controllers. For this purpose we ignore the
format `:php` through a "route constraint".
Some other requests are still passed to the
`Goldencobra::ArticlesController` though. These might have the format
`:xml` or something similar, that we generally accept. Now we have to
make sure, that we send the correct response if the resource isn't
found.

For that purpose we use `format.any { head(status: 404) }` inside our
`redirect_to_404` method. Up until now there was only a `respond_to`
block for the `format.html` case. If the request has the format `:html`
we continue to serve our (custom) 404 article/page. If the request is of
a different format, we kept receiving `ActionController::UnknownFormat:
ActionController::UnknownFormat` errors. This new format block will
prevent this error from happening. Just respond with a "HEAD 404"
response is enough for these cases.