Only create tables and indexes if not present
The option `force: :cascade` deletes all tables and indexes and recreates them from sratch. This leads to empty dbs after an upgrade form Goldencobra 1.x to 2.x, because this migration is new to all 1.x projects. It would get run, and the contents of the db would be evaporated. This is …bad!

That's why we only use the contents of this migration if the tables and indexed don't actually exist. The next few migrations in Goldencobra are needed, as no 1.x project would have these. But they don't recreate the tables, so they are cool.