Shorten cache key generation for over-view tree
The generation of the cache key lead to an `ENAMETOOLONG` error:
```
Errno::ENAMETOOLONG - File name too long @ rb_file_s_stat
```

If we use the the `#hash` method of Strings we can prevent this from
happening. The hashed strings are still random enough for the cache to
work properly.

There is a minor concern with this method:

> The hash value for an object may not be identical across invocations
or implementations of Ruby. If you need a stable identifier across Ruby
invocations and implementations you will need to generate one with a
custom method.

But in practice this shouldn't really be an issue. Even if the cache is
missed because the key generated does not exist in this invocation, the
cache contents will just be regenerated.

Fixes #74