Bugfix for empty custom tag settings
If there were no custom tag settings defined as Goldencobra::Setting,
the method to merge these custom settings with the given article's
settings would try to merge with `nil`. This lead to an error, as you
can't merge a Hash with nil. The method for returning the custom tags
needs to return an empty hash, if no custom tags are defined.