Add service for cleaning up unused settings
There are settings listed in the backend and nobody knows, if they
are used or not. A lot of settings were changed to locales and
others are not used completely.
So they do not need to be filled with values.

Solution:

Unused settings should be removed from the database.
Therefore methods are introduced to check the
database against a yml-file and to check the code
for presence of setting keys.
If a setting in the database cannot be found in
the yml-file, it is unused and can be removed.
If a setting is not used in the code, its key will
be logged to the console and can be removed
from the yml-file.

Further:

You also have to clean the website settings from dummy values,
because otherwise they will be rendered on pages.
This removes unnecessary values from the settings yml-file.

GCZ-70