Change default setting for geocoding to false
There is a google quota which is reached fastly without api key. We are using geocoding without api key.

Usage limits:
https://developers.google.com/maps/faq?hl=de#usage-limits

In most cases the apps do not need geocoding, so this commits changes the default value of the geocoding setting from true to false.

For new apps it is important to think about geocoding from now on. If it is needed, the setting needs to be changed to true.

GC-163