Move column preview as first column in admin/upload index
The preview is the most important part in this table, so
let it be the first column. In addition it is rendered
a bit bigger with size `:thumb` instead of `:mini` and
it links to the preview single page on an upload.

Further added some missing locale keys and updated translations.

GC-8