Prevent errors on index and on create for uploads
Sometimes uploads cannot be created properly because of imagemagick problems or other reasons. The
create action needs to redirect(:back) otherwise there is an 500 error screen.
Also the index needs to take into account that some uploads might be faulty. You need to display
them in the index nonetheless to be able to delete them.