Fix online offline setting for articles
The slug was used wrongly instead of the id
of an article to find it. If an article
was not found, there was no catching case.

This commit adds a correct use of id and
a case where article is not found per
member action for setting it online or offline.