Remove Bootstrap 3 gem from gemspec
Gemspec was specifying bootstrap 3 gem. The problem with that was that any host application which uses Goldencobra
cannot use any other bootstrap version via the gem.
This commit will remove bootstrap 3 gem from gemspec. By that it is possible to e.g use the bootstrap 4 gem with GC
in any host application, together with the Goldencobra Engine.